Place inventory logic in it's own module
Split the inventory logic out to it's own module
to make it more manageable.
